Use rigtorp/SPSCQueue for MessagePipe
