+++ src/engine/effects/engineeffect.cpp
@@ -104,7 +104,7 @@ bool EngineEffect::processEffectsRequest(EffectsRequest& message,
             }
 
             response.success = true;
-            pResponsePipe->writeMessage(&response);
+            pResponsePipe->writeMessage(response);
             return true;
             break;
         case EffectsRequest::SET_PARAMETER_PARAMETERS:
@@ -128,7 +128,7 @@ bool EngineEffect::processEffectsRequest(EffectsRequest& message,
                 response.success = false;
                 response.status = EffectsResponse::NO_SUCH_PARAMETER;
             }
-            pResponsePipe->writeMessage(&response);
+            pResponsePipe->writeMessage(response);
             return true;
         default:
             break;
