+++ src/engine/effects/engineeffectchain.cpp
@@ -141,7 +141,7 @@ bool EngineEffectChain::processEffectsRequest(EffectsRequest& message,
         default:
             return false;
     }
-    pResponsePipe->writeMessage(&response);
+    pResponsePipe->writeMessage(response);
     return true;
 }
