+++ src/util/fifo.h
@@ -1,14 +1,10 @@
 #ifndef FIFO_H
 #define FIFO_H
 
-#include <QtDebug>
-#include <QScopedPointer>
-#include <QSharedPointer>
-
 #include "pa_ringbuffer.h"
+
 #include "util/class.h"
 #include "util/math.h"
-#include "util/reference.h"
 
 template <class DataType>
 class FIFO {
@@ -72,103 +68,6 @@ class FIFO {
     DataType* m_data;
     PaUtilRingBuffer m_ringBuffer;
     DISALLOW_COPY_AND_ASSIGN(FIFO<DataType>);
-};
-
-// MessagePipe represents one side of a TwoWayMessagePipe. The direction of the
-// pipe is with respect to the owner so sender and receiver are
-// perspective-dependent.
-template <class SenderMessageType, class ReceiverMessageType>
-class MessagePipe {
-  public:
-    MessagePipe(FIFO<SenderMessageType>& receiver_messages,
-                FIFO<ReceiverMessageType>& sender_messages,
-                BaseReferenceHolder* pTwoWayMessagePipeReference)
-            : m_receiver_messages(receiver_messages),
-              m_sender_messages(sender_messages),
-              m_pTwoWayMessagePipeReference(pTwoWayMessagePipeReference) {
-    }
-
-    // Returns the number of ReceiverMessageType messages waiting to be read by
-    // the receiver. Non-blocking.
-    inline int messageCount() const {
-        return m_sender_messages.readAvailable();
-    }
-
-    // Try to read read a ReceiverMessageType written by the receiver
-    // addressed to the sender. Non-blocking.
-    inline bool readMessage(ReceiverMessageType* message) {
-        return m_sender_messages.read(message, 1) > 0;
-    }
-
-    // Writes a message from the 'message' array to the receiver
-    // and returns true on successess.
-    inline bool writeMessage(const SenderMessageType* message) {
-        int result = m_receiver_messages.write(message, 1);
-        return result;
-    }
-
-  private:
-    FIFO<SenderMessageType>& m_receiver_messages;
-    FIFO<ReceiverMessageType>& m_sender_messages;
-    QScopedPointer<BaseReferenceHolder> m_pTwoWayMessagePipeReference;
-
-#define COMMA ,
-    DISALLOW_COPY_AND_ASSIGN(MessagePipe<SenderMessageType COMMA ReceiverMessageType>);
-#undef COMMA
-};
-
-// TwoWayMessagePipe is a bare-bones wrapper around the above FIFO class that
-// facilitates non-blocking two-way communication. To keep terminology clear,
-// there are two sides to the message pipe, the sender side and the receiver
-// side. The non-blocking aspect of the underlying FIFO class requires that the
-// sender methods and target methods each only be called from a single thread
-// The most common use-case of this class is sending and receiving messages
-// with the callback thread without the callback thread blocking.
-//
-// This class is an implementation detail and cannot be instantiated
-// directly. Use makeTwoWayMessagePipe(...) to create a two-way pipe.
-template <class SenderMessageType, class ReceiverMessageType>
-class TwoWayMessagePipe {
-  public:
-    // Creates a TwoWayMessagePipe with SenderMessageType and
-    // ReceiverMessageType as the message types. Returns a pair of MessagePipes,
-    // the first is the sender's pipe (sends SenderMessageType and receives
-    // ReceiverMessageType messages) and the second is the receiver's pipe
-    // (sends ReceiverMessageType and receives SenderMessageType messages).
-    static QPair<MessagePipe<SenderMessageType, ReceiverMessageType>*,
-                 MessagePipe<ReceiverMessageType, SenderMessageType>*> makeTwoWayMessagePipe(
-                     int sender_fifo_size,
-                     int receiver_fifo_size) {
-        QSharedPointer<TwoWayMessagePipe<SenderMessageType, ReceiverMessageType> > pipe(
-            new TwoWayMessagePipe<SenderMessageType, ReceiverMessageType>(
-                sender_fifo_size, receiver_fifo_size));
-
-        return QPair<MessagePipe<SenderMessageType, ReceiverMessageType>*,
-                     MessagePipe<ReceiverMessageType, SenderMessageType>*>(
-                         new MessagePipe<SenderMessageType, ReceiverMessageType>(
-                             pipe->m_receiver_messages, pipe->m_sender_messages,
-                             new ReferenceHolder<TwoWayMessagePipe<SenderMessageType, ReceiverMessageType> >(pipe)),
-                         new MessagePipe<ReceiverMessageType, SenderMessageType>(
-                             pipe->m_sender_messages, pipe->m_receiver_messages,
-                             new ReferenceHolder<TwoWayMessagePipe<SenderMessageType, ReceiverMessageType> >(pipe)));
-    }
-
-  private:
-    TwoWayMessagePipe(int sender_fifo_size, int receiver_fifo_size)
-            : m_receiver_messages(receiver_fifo_size),
-              m_sender_messages(sender_fifo_size) {
-    }
-
-    // Messages waiting to be delivered to the receiver.
-    FIFO<SenderMessageType> m_receiver_messages;
-    // Messages waiting to be delivered to the sender.
-    FIFO<ReceiverMessageType> m_sender_messages;
-
-    // This #define is because the macro gets confused by the template
-    // parameters.
-#define COMMA ,
-    DISALLOW_COPY_AND_ASSIGN(TwoWayMessagePipe<SenderMessageType COMMA ReceiverMessageType>);
-#undef COMMA
 };
 
 #endif /* FIFO_H */
